refactor theme toggle from App to Layout component
