+++ src/contexts/theme.js
@@ -1,12 +1,18 @@
 import {createContext} from 'react'
 
 const BROWSER_DARK = '(prefers-color-scheme: dark)'
-
 const themes = ['dark', 'light']
-const userBrowserTheme = window.matchMedia(BROWSER_DARK).matches ? 'dark' : 'light'
+
+const checkBrowserThemeDark = () => window
+		.matchMedia(BROWSER_DARK)
+		.matches
+
+const userBrowserTheme = checkBrowserThemeDark() ? themes[0] : themes[1]
 
 // If user changes preference, toggle the class.
 /* window.matchMedia(DARK).addEventListener('change', ({matches}) => {}) */
+
+console.log('userBrowserTheme', userBrowserTheme)
 
 const themeContextDefault = {
 	theme: userBrowserTheme,
@@ -19,5 +25,6 @@ const ThemeContext = createContext(themeContextDefault)
 
 export {
 	ThemeContext,
-	themeContextDefault
+	themeContextDefault,
+	checkBrowserThemeDark,
 }
